Add import in failing to compile Source file
For some unknown reason this import is not required when compiling
inside sbt. This is likely due to the way these generated sources and
compiled and then put into the classpath of scio-extras-test.

In bloop, the .class associated with the generated avro files are
available in the scio-extras compile target.